feat(memo): implement MemoEditor and MemoEditorEmpty components; enhance MemoTab with MemoSidebar integration
